Fixed unit test for floating point config variables.
See issue #101

+++ multicore_engine_tests/src/config/config_test.cpp
@@ -48,10 +48,10 @@ TEST(config_config_store, load_config_float) {
 	std::stringstream ustr;
 	ustr.str("test.name1=135.79\n");
 	config_store cs(ustr, dstr, [](config_store::config_storer&) {});
-	auto test_name1 = cs.resolve<int>("test.name1");
+	auto test_name1 = cs.resolve<float>("test.name1");
 	ASSERT_FLOAT_EQ(135.79, test_name1->value());
 	ASSERT_FALSE(test_name1->dirty());
-	auto test_name2 = cs.resolve<int>("test.name2");
+	auto test_name2 = cs.resolve<float>("test.name2");
 	ASSERT_FLOAT_EQ(67.89, test_name2->value());
 	ASSERT_FALSE(test_name2->dirty());
 }
